Module hopr_internal_types::channels

Implements types related to HOPR payment channels.

Structs

ChannelEntry

Overall description of a channel

Ticket

Contains the overall description of a ticket with a signature

Enums

ChannelChange

Enumerates possible changes on a channel entry update

ChannelDirection

Describes a direction of node’s own channel. The direction of a channel that is not own is undefined.

ChannelStatus

Describes status of a channel

Functions

f64_to_win_prob

Encodes [0.0f64, 1.0f64] to [0x00000000000000, 0xffffffffffffff]

generate_channel_id

win_prob_to_f64

Decodes [0x00000000000000, 0xffffffffffffff] to [0.0f64, 1.0f64]

Type Aliases

EncodedWinProb

Winning probability encoded in 7-byte representation