hopr_internal_types

Module path

Implements types for path selection and construction.

Structs

ChainPath

Represents a NonEmptyPath that completely specifies a route using Addresses.

NoPathSelector

A path selector that does not resolve any path, always returns PathError::PathNotFound.

TransportPath

A NonEmptyPath that can be used to route packets using [OffchainPublicKey].

ValidatedPath

Represents NonEmptyPath that has been resolved and validated.

Traits

NonEmptyPath

A Path that is guaranteed to have at least one hop - the destination.

Path

Base implementation of an abstract path.

PathAddressResolver

Allows resolution of [OffchainPublicKey] for a given [Address] or vice versa and retrieval of ChannelEntry based on the parties.

PathSelector

Trait for implementing a custom path selection algorithm from the channel graph.

Type Aliases

ChannelPath