Module hopr_internal_types::protocol

Types related to internal HOPR protocol logic.

Structs

ApplicationData

Represents the received decrypted packet carrying the application-layer data.

TagBloomFilter

Bloom filter for packet tags to detect packet replays.

Constants

DEFAULT_APPLICATION_TAG

Represent a default application tag if none is specified in send_packet.

INTERMEDIATE_HOPS

Number of intermediate hops: 3 relayers and 1 destination

PAYLOAD_SIZE

Maximum size of the packet payload in bytes.

TICKET_WIN_PROB

Fixed ticket winning probability

Type Aliases

Tag

Tags are currently 16-bit unsigned integers