hopr_internal_types

Module channels

Implements types related to HOPR payment channels.

Structs

ChannelEntry

Overall description of a channel

ChannelParties

An immutable pair of addresses representing parties of a channel.

CorruptedChannelEntry

A wrapper around ChannelId representing a Channel that is corrupted.

SrcDstPair

A pair of source and destination addresses representing a channel.

Enums

ChannelChange

Lists possible changes on a channel entry update

ChannelDirection

Describes a direction of node’s own channel. The direction of a channel that is not own is undefined.

ChannelStatus

Describes status of a channel

ChannelStatusDiscriminants

Functions

generate_channel_id

Generates channel ID hash from source and destination addresses.

Type Aliases

ChannelId

Alias for the Hash representing a channel ID.