hopr_internal_types
Module protocol

Types related to internal HOPR protocol logic.

Structs

Acknowledgement

Represents packet acknowledgement

ApplicationData

Represents the received decrypted packet carrying the application-layer data.

TagBloomFilter

Bloom filter for packet tags to detect packet replays.

Enums

PendingAcknowledgement

Contains either unacknowledged ticket if we’re waiting for the acknowledgement as a relayer or information if we wait for the acknowledgement as a sender.

Constants

DEFAULT_APPLICATION_TAG

Represent a default application tag if none is specified in send_packet.

DEFAULT_MAXIMUM_INCOMING_TICKET_WIN_PROB

Default maximum incoming ticket winning probability, above which tickets will not be accepted due to privacy.

DEFAULT_MINIMUM_INCOMING_TICKET_WIN_PROB

Default required minimum incoming ticket winning probability

DEFAULT_OUTGOING_TICKET_WIN_PROB

Default ticket winning probability that will be printed on outgoing tickets

INTERMEDIATE_HOPS

Number of intermediate hops: 3 relayers and 1 destination

LOWEST_POSSIBLE_WINNING_PROB

The lowest possible ticket winning probability due to SC representation limit.

PAYLOAD_SIZE

Maximum size of the packet payload in bytes.

Type Aliases

Tag

Tags are currently 16-bit unsigned integers