hopr_internal_types

Module protocol

Types related to internal HOPR protocol logic.

Structs

Acknowledgement

Unverified packet acknowledgement.

VerifiedAcknowledgement

Represents packet acknowledgement whose signature has been already verified.

Enums

PendingAcknowledgement

Contains either unacknowledged ticket if we’re waiting for the acknowledgement as a relayer or information if we wait for the acknowledgement as a sender.

Constants

DEFAULT_MAXIMUM_INCOMING_TICKET_WIN_PROB

Default maximum incoming ticket winning probability, above which tickets will not be accepted due to privacy.

DEFAULT_MINIMUM_INCOMING_TICKET_WIN_PROB

Default required minimum incoming ticket winning probability

INTERMEDIATE_HOPS

Number of intermediate hops: 3 relayers and 1 destination

MIN_BATCH_SIZE

Minimum batch size of acknowledgements to use EdDSA batch verification algorithm.

Type Aliases

HoprPseudonym

Alias for the Pseudonym used in the HOPR protocol.