hopr_internal_types

Module tickets

Implements types for tickets.

Structs

AcknowledgedTicket

Contains acknowledgment information and the respective ticket

RedeemableTicket

Represents a winning ticket that can be successfully redeemed on-chain.

Ticket

Contains the overall description of a ticket with a signature.

TicketBuilder

Builder for Ticket and VerifiedTicket.

TransferableWinningTicket

Represents a ticket that could be transferred over the wire and independently verified again by the other party.

UnacknowledgedTicket

Represents a VerifiedTicket with an unknown other part of the [HalfKey]. Once the other [HalfKey] is known (forming a [Response]), it can be acknowledged.

VerifiedTicket

Holds a ticket that has been already verified. This structure guarantees that Ticket::get_hash() of VerifiedTicket::verified_ticket() is always equal to VerifiedTicket::verified_hash

WinningProbability

Represents a ticket winning probability.

Enums

AcknowledgedTicketStatus

Status of the acknowledged ticket.

Constants

REDEEM_CALL_SELECTOR

Define the selector for the redeemTicketCall to avoid importing the entire hopr-bindings crate for one single constant. This value should be updated with the function interface changes.

Type Aliases

EncodedWinProb

Winning probability encoded in 7-byte representation