hopr_internal_types::tickets

Struct TicketBuilder

pub struct TicketBuilder { /* private fields */ }

Builder for the Ticket and VerifiedTicket.

A new builder is created via TicketBuilder::default or TicketBuilder::zero_hop.

Input validation is performed upon calling TicketBuilder::build, TicketBuilder::build_signed and TicketBuilder::build_verified.

Implementations

impl TicketBuilder

pub fn zero_hop() -> Self

Initializes the builder for a zero-hop ticket.

pub fn counterparty<A: Into<Address>>(self, counterparty: A) -> Self

Sets the counterparty (recipient) of the ticket.

pub fn amount<T: Into<U256>>(self, amount: T) -> Self

Sets the ticket amount. This or TicketBuilder::balance must be set and be less or equal to 10^25.

pub fn balance(self, balance: HoprBalance) -> Self

Sets the ticket amount as HOPR balance. This or TicketBuilder::amount must be set and be less or equal to 10^25.

pub fn index(self, index: u64) -> Self

Sets the ticket index. Must be less or equal to 2^48. Defaults to 0.

pub fn channel_epoch(self, channel_epoch: u32) -> Self

Sets the channel epoch. Must be less or equal to 2^24. Defaults to 1.

pub fn win_prob(self, win_prob: WinningProbability) -> Self

Sets the ticket winning probability. Defaults to 1.0

pub fn challenge(self, challenge: Challenge) -> Self

Sets the [Challenge] for the Proof of Relay, converting it to [EthereumChallenge] first.

Either this method or [Ticket::eth_challenge] must be called.

pub fn eth_challenge(self, challenge: EthereumChallenge) -> Self

Sets the [EthereumChallenge] for the Proof of Relay. Either this method or Ticket::challenge must be called.

pub fn signature(self, signature: Signature) -> Self

Set the signature of this ticket. Defaults to None.

pub fn build(self) -> Result<Ticket>

Verifies all inputs and builds the Ticket. This does not perform signature verification if a signature was set.

pub fn build_signed( self, signer: &ChainKeypair, domain_separator: &Hash, ) -> Result<VerifiedTicket>

Validates all inputs and builds the VerifiedTicket by signing the ticket data with the given key. Fails if signature was previously set.

pub fn build_verified(self, hash: Hash) -> Result<VerifiedTicket>

Validates all inputs and builds the VerifiedTicket by assuming the previously set signature is valid and belongs to the given ticket hash.

It does not check whether hash matches the input data nor that the signature verifies the given hash.